requirements.txt
pip==21.3.1
python-telegram-bot==13.13
psycopg2==2.9.3
certifi==2022.6.15
pytz==2022.1
tzdata==2022.1
setuptools==60.2.0
click==8.1.3
tornado==6.2
importlib-metadata==4.12.0
cachetools==4.2.2
zipp==3.8.1
APScheduler==3.6.3
tzlocal==2.1
idna==3.3
Jinja2==3.1.2
requests==2.28.1
python-dateutil==2.8.2
Flask==2.2.2
Werkzeug==2.2.2
itsdangerous==2.1.2
charset-normalizer==2.1.1
urllib3==1.26.12
numpy==1.23.1
six==1.16.0
schedule==1.1.0
wheel==0.37.1
colorama==0.4.5
pytz-deprecation-shim==0.1.0.post0